third_party/blink/web_tests/css3/calc/cssom-expected.txt

This tests calc() and the CSSOM

10px => calc(10px)
10px + 15px => calc(25px)
100% => calc(100%)
100% - 10px => calc(100% - 10px)
10px + 10px * 5 => calc(60px)
5px + 2em + 6in => calc(5px + 2em + 6in)
100% - 10px / 2 => calc(100% - 5px)
1px + 2em - 3rem + 4in => calc((1px + 2em) - 3rem + 4in)
100px * (1 + 2 * 3 - 4 / 5) => calc(620px)
(100px) + 200px => calc(300px)
((((((((((100px)))))))))) => calc(100px)
flimstix => calc(100px)
